+++ src/Security/RegistrationController.php
@@ -1,0 +1,32 @@
+<?php
+
+
+namespace App\Security;
+
+
+use App\Entity\User;
+use App\Form\RegistrationFormType;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+
+class RegistrationController extends AbstractController
+{
+    /**
+     * @Route("/inscription", name="security_registration")
+     */
+    public function registration(Request $request, EntityManagerInterface $entityManager){
+
+        $user = new User();
+        $form = $this->createForm(RegistrationFormType::class, $user);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($user);
+            $entityManager->flush();
+        }
+        return $this->render('security/registration.html.twig', [
+            'form' => $form->createView()
+        ]);
+    }
+}
