+++ templates/user/component/navbar.html.twig
@@ -44,6 +44,7 @@
             </li>
             <li class="nav-item">
                 <a class="nav-link " href="{{ path("app_login") }}">login</a>
+
             </li>
             <li class="nav-item">
                 <a class="nav-link " href="{{ path("app_logout") }}">Logout</a>
