+++ templates/user/pages/article.html.twig
@@ -1,7 +1,7 @@
 {% extends 'base.html.twig' %}
 
 
-{% block style %}
+{% block styles %}
     <link rel="stylesheet" href="{{ asset('stylesheet/user_style/article_style.css') }}">
 {% endblock %}
 
@@ -25,6 +25,8 @@
             <div class="mt-5 ml-5 mr-5 content">{{ article.content }}</div>
         </div>
     </div>
+    <h2>Ajouter un commentaire</h2>
 
+    {#{{ form(form) }}#}
 {%  endblock %}
